Cambiando de lugar un método
